Implemented basic camera function and wrappers
Implemented takepic() in camera.c that will build a raspistill command to take a picture
using the raspberry pi camera module v2. This is currently working, although a more
elegant solution should be implemented for later use. For a start, two wrapper functions
were created to allow for the use of the Grovepi C++ library in C, since all the functions
there are contained within a namespace with some utilizing exceptions. Further work will
have to be done to ensure no exceptions are not allowed to propigate into the base C code.