Major Changes, Implemented Basic System
Implemented the basic verison of the birdwatcher system, for now any
pictures and enviornmental data are stored in the filesystem under the
directory where the executable is located. Each day will have its own
folder for storing pictures and a logfile for storing enviornmental
data. Right now time is measured to make a reading every fifteen
minutes. In the future, it would be better to utilize an alarm or a
seperate thread for the logging of this information, especially as it
takes a long time to take a picture.