Update License
Re-licensed under the MIT License. The project license has been updated to reflect its current state. Due to unfortunate circumstances I will be finishing the project by myself. As all the code up to this point was written by me, the license and readme are the only things that will be changed as a result