Updated Method for Image Capture
Pictures will now be taken by sending a USR1 signal to a running
raspistill process instead of running raspistill for each picture. A
consequence of this is the raspistill process needs to be terminated
before the program exits regardless of the exit reason. Some output from
an apparent failed kill(1) execution that appears when interuppting
the program will have to be addressed. As well as removing the temporary
image file once it is no longer needed.