Updated Environmental Logic, Minor Bug Fixes
Environmental readings now occur via an alarm instead of checking if 15
minutes has passed via getting the system time. This alarm is reset
after it goes off, so the next reading can be taken in fifteen minutes
Updated the signal handler to also handle SIGALRM. Now the handler will
set one of two flags when a signal is recieved from the Kernel.

All output from the raspistill subprocess is now redirected to
/dev/null. This was done by redirecting stdout to /dev/null and then
stderr to stdout. This prevents error messages from appearing when the
raspistill process is killed. The cleanup function might not be needed
as this process seems to end when the program is terminated.