Added Basic Thingspeak integration
Basic ThingSpeak functionality has been added via libcurl. This is done
via an http POST request. A utility function has been added to read the
API key for thingspeak from a file. This file will not be tracked in the
repository.

Todo: The use of sleep() after taking a picture with the camera will
interfere with the 15 minute alarms set to keep track of when
environmental readings should be taken. This needs to be fixed in the
future. In addition, an api for picture storage needs to be implemented.