Updated Timing Methods
The timer used to determine when an environmental reading should be
taken is now based off of a USR1 signal instead of SIGALRM so that
sleep() can be safely used. In addition, another timer was added using
a USR2 signal to time out the camera for a minute so it won't take rapid
fire pictures.

TODO: add some networking component to store the pictures online.
Cleanup existing code, especially in the main function.