Added Imgur Integration
Pictures captured will now be uploaded to a public imgur album under the
username raspibirdwatcher. Like the thingspeak integration from before,
this upload is performed via an http POST request using libcurl. In
future, a description and name should be applied to the picture when it
is uploaded.

Todo: The responce from the Imgur API should either be output into a
file or discarded, instead of printed to stdout.