Added Description to Imgur Upload
A description indicating the date and time a picture was taken has been
added to the Imgur POST request. In addition, indication has been made
when the upload begins and completes.

Unfortunately, I had left the oauth key visible in a source file in the
previous commit. This key has been removed, and a replacement key
obtained.